+++ src/constants/tools.ts
@@ -4,7 +4,6 @@ export const TELEGRAM_SEND_TOOLS = new Set([
   "telegram_send_voice",
   "telegram_send_sticker",
   "telegram_send_photo",
-  "telegram_send_rich_message",
   "telegram_forward_message",
   "telegram_quote_reply",
   "telegram_send_buttons",